Remove useless permissions
fcitx was introduced in https://github.com/flathub/com.discordapp.Discord/pull/63 due to a fcitx change which should no longer be needed.
The env variable for ubuntu:GNOME was done for experiments regarding application badges. I never managed to make it work and I don't think this will help with that so it's removed