Disable forced updates before the first startup (#265)
Currently the `disable-breaking-updates.py` script disables updates only
if the `settings.json` file already exists in the Discord config
directory.  This renders the flatpak essentially useless for new users
in the short intervals between a new update being released by Discord
and the new version being available on Flathub, see #263, #262 and #252
for some recent examples.

This change makes the script write a dummy `settings.json` file with
updates disabled if it doesn't already exist, hopefully eliminating this
kind of issues.