Build socat with most features disabled
socat offers a plethora of features, but we don't need most of them.

This configures socat with all features disabled, EXCEPT these:

- --disable-unix
- --disable-ip4 [if we include this one, the build breaks :(]
- --disable-listen
- --disable-retry

I wish there was a way to have every feature disabled by default so
that we would only enable the ones we need, but that's not possible.

This change also removes two unused installed executables (filan and
procan), three scripts (socat-{chain,broker,mux}.sh), and a manpage.