Update Discord to 0.0.87
This updates Discord from 0.0.74 to 0.0.87.

This big jump in version is due to a deliberate choice not to update,
to prevent a breaking change introduced with Electron 32.2, which ended
up breaking the file chooser portal for most people. [1]

That change began requiring version 1.20.0 of xdg-desktop-portal, which
at the time was not even a stable release. Today, 1.20.0 is stable, but
it will take time before it's finally available for users of
non-rolling-release distributions.

Please note that Discord 0.0.87 is still affected by the bug above, but
we're much closer to seeing this bug get fixed now: The breaking change
was reverted in Electron versions 32.3 and 34.0+, with the fix for
version 33 still being worked on (at the time of writing this). [2]

And, unluckily for us, 0.0.87 includes Electron 33, which does not have
the fix yet. Still, the bump to Electron 34+ should, hopefully, not take
much longer...

The main reason for pushing this update now is that Discord 0.0.80+ added
proper screen-sharing support on Wayland, so users expect us to update as
well. [3]

Originally, I was also going to enable Wayland by default with this
update, but that experiment exposed another Electron/Chromium bug that
affects GNOME (one of the most used desktop environments), so I had to
revert it. The bug in question prevents users from normally resizing
Discord's window with their mouse. [4]

Despite that, I think we're getting closer to enabling Wayland by default
in the near future. [5]

Lastly, this change also re-enables Discord update checking from f-e-d-c.

References:

[1] https://github.com/electron/electron/issues/43819
[2] https://github.com/electron/electron/pull/45180
[3] https://github.com/flathub/com.discordapp.Discord/issues/380#issuecomment-2593001344
[4] https://github.com/electron/electron/issues/44543
[5] https://nickdiego.dev/blog/chromium-ozone-wayland-the-last-mile-stretch/