Add --disable-gpu-memory-buffer-video-frames on Wayland
While I was testing the newly-added screen-sharing Discord feature for
Wayland, my terminal got spammed by the following errors:

[65:0219/130307.530052:ERROR:gbm_pixmap_wayland.cc(82)] Cannot create bo with format= YUV_420_BIPLANAR and usage=SCANOUT_CPU_READ_WRITE
[65:0219/130307.530052:ERROR:gpu_channel.cc(503)] Buffer Handle is null.

These errors only appeared when sharing the entire screen and when
Discord's main window was in the foreground.

It appears to be a real Chromium bug [1] that was fixed in M-131 [2].

However, Discord 0.0.87 still uses Electron 33, which in turn uses
Chromium M-130, so that explains why we're still affected.

This is just a workaround, and we will likely remove this launch option
in future Discord updates.

[1] https://issues.chromium.org/issues/331796411
[2] https://chromium.googlesource.com/chromium/src.git/+/000064171db044f98d1c9e025e16007d8269e26a%5E!/